Update gh action and drop travis
