--- .travis.yml
@@ -1,13 +1,0 @@
-language: python
-python:
-    - "2.7"
-env:
-    - PGVERSION=9.1 CKANVERSION=master
-    - PGVERSION=9.1 CKANVERSION=2.3
-install:
-    - bash bin/travis-build.bash
-    - pip install coveralls
-script: sh bin/travis-run.sh
-after_success:
-    - coveralls
-sudo: required
